[add] - creating a file sort with number on the idea of bitmap with
