+++ BitSorting/BitSorting.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30011.22
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BitSorting", "BitSorting\BitSorting.vcxproj", "{49036D08-81AA-4FBA-92BA-E782EF4557C0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{49036D08-81AA-4FBA-92BA-E782EF4557C0}.Debug|x64.ActiveCfg = Debug|x64
+		{49036D08-81AA-4FBA-92BA-E782EF4557C0}.Debug|x64.Build.0 = Debug|x64
+		{49036D08-81AA-4FBA-92BA-E782EF4557C0}.Debug|x86.ActiveCfg = Debug|Win32
+		{49036D08-81AA-4FBA-92BA-E782EF4557C0}.Debug|x86.Build.0 = Debug|Win32
+		{49036D08-81AA-4FBA-92BA-E782EF4557C0}.Release|x64.ActiveCfg = Release|x64
+		{49036D08-81AA-4FBA-92BA-E782EF4557C0}.Release|x64.Build.0 = Release|x64
+		{49036D08-81AA-4FBA-92BA-E782EF4557C0}.Release|x86.ActiveCfg = Release|Win32
+		{49036D08-81AA-4FBA-92BA-E782EF4557C0}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {504BBBFC-C15F-43AE-B17B-CE685D2B873C}
+	EndGlobalSection
+EndGlobal
